fix(rocsparse): disambiguate gebsrmm grid.y min() for ILP64 builds (AISPARSE-666)
With -DBUILD_ROCSPARSE_ILP64=ON the block-column count is 64-bit, so
rocsparse::min((n - 1) / N_ + 1, 65535) was ambiguous between the int32_t and
int64_t min overloads. Cast both operands to int64_t at the four gebsrmm grid.y
clamps (general, large, large_ext, small).

+++ projects/rocsparse/library/src/level3/rocsparse_gebsrmm_template_general.cpp
@@ -108,7 +108,9 @@ namespace rocsparse
 
         hipStream_t stream = handle->stream;
 
-        dim3 gebsrmm_blocks((mb - 1) / 1 + 1, rocsparse::min((n - 1) / 32 + 1, 65535));
+        dim3 gebsrmm_blocks(
+            (mb - 1) / 1 + 1,
+            rocsparse::min(static_cast<int64_t>((n - 1) / 32 + 1), static_cast<int64_t>(65535)));
         dim3 gebsrmm_threads(32, 32, 1);
         RETURN_IF_HIPLAUNCHKERNELGGL_ERROR((rocsparse::gebsrmm_general_blockdim_kernel<32, 32>),
                                            gebsrmm_blocks,
